revert: "Merge pull request #3 from Idrinth/add-multi-translation-support"
It's not possible to get the game language before kDataLoaded. Since we're hooking into the file stream function of skyrim which runs before kDataLoaded, we need it early to get the language overwrite files. One option would be to remove auto and let the user enter it. Another would be to manually parse Skyrim.ini and Skyrimcustom.ini.